Move Occupancy to TripKitDomain
